+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.44] - 2023-04-12
+### Added
+* `PoolSystem` and `IPoolable<T>` which serves as a similar framework to `ListPool`
+### Changed
+* `ListPool` directory to Pooling
+* `ListPool`.`Release` to `Add`
+
 ## [0.0.43] - 2023-04-06
 ### Added
 * reference NOTE comment to `ListPool`